requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
gunicorn==21.2.0
python-dotenv==1.0.1
pytz==2025.1
python-barcode==0.15.1
Pillow==10.2.0
